Some what working version for VoiceMOS
config/test.json is the config to use. It perturbs (modifies SNR(for now)) an input signal from a path given in the dataframe (from csv files). Spectrograms are created for the reference and perturbed signals and then the NSIM is calculated for each perturbation. The final NSIM matrix for each signal is written back to the dataframe at the end

NOTE: HIGHLY EXPERIMENTAL BRANCH, LOTS OF MESSY CODE AND BROKEN/DISABLED FEATURES